I haven't yet been able to figure out how to afford $500+ month plaid liabilities product without going into debt and I don't want to complicate the UI or back end so authoritative debts from a financial institution don't have a calculated payoff.
Instead what I have come to as a compromise is observing the interest rate going down in real-time. It sucks, because the financial institution doesn't create an amortization chart based on your current payment. Instead the financial institution does an amortization chart in the worst case scenario of your minimum payment. It's not right, but I don't have the user base to change it. If I complicate the UI then I with user-defined interest rates on top of current loans from a financial institution the product becomes too complex for anyone to use.